stretching improvements (from C. Donaldson)
Note: this is a very unsafe version, i.e. the strectching must be done only once on the complete archive. If ran in cron, it'll overwrite the previous results.
Future (final) version will include the export of the whole corr-coeff array to be able to mimic standard use of stretching (like in CS or TR papers).