Fix use of tables prefix in database creation
msnoise_table_def.py was defining the table names and reading the prefix
from the db.ini file at compilation time, right after the interpreter
starts (as the code was at module level, in the class declaration).

As a consequence, on a new project (with no db.ini file), the prefix
chosen by the user was added to the db.ini file but ignored when
creating the tables, which were always created without any prefix. When
re-initialising the database (when a db.ini already exists), only the
prefix written in db.ini during the last run of 'msnoise db init' was
used, and again, the prefix entered by the user was ignored and simply
stored in db.ini.

As SQLAlchemy Declarative system sets up the classes mapped to
relational database tables as soon as they are defined, this definition
has to be postponed to run-time to allow the table names to use a prefix
entered by the user.

The changes introduced by this commit move the definition of the mapped
classes to a function to allow the use of a prefix entered before by the
user (when initialising a new database) or read from the db.ini file
(when using an existing database).

Modules using these mapped classes are updated to call the function
defining the mapped class instead of directly importing the classes.

Additionally, the s000installer.py file is reworked for better
readability and maintainability:

 - indices are created in their own function

 - user input for database and password choice somewhat improved

 - improved readability and maintainability of code handling default
   values in user choices

 - for security, no default password is provided (to avoid weak default
   installation)