Fix database table prefix regression
Commit 6cefc25 introduced a regression for plugins on database tables
creation.  This commit simplifies the changes made by 6cefc25 and allows
to retain compatibility with existing plugins.

With this commit, msnoise_table_def.py by default defines module classes
mapped to relational database tables using the prefix defined in the
db.ini file (or an empty prefix if the file is missing or not readable).
The only case where this behaviour is not desirable is when initialising
a new database, from s000installer.py.  In this case, we simply redefine
the mapped classes using the new suffix to use.