Better handling of a missing db.ini
As the db.ini is looked up through api.connect() very early at runtime,
even before any provided subcommand is even taken into account, db.ini
can legitimately not exist (as it is the case when running 'msnoise db
init').  api.read_database_inifile() should therefore not do a
sys.exit() when the file is not found, as an exit is not always wanted.

This was previously resolved by a very broad 'except:' catching the
SystemExit exception emitted by sys.exit(), but this would mask any
other exceptions and make debugging potential problem difficult.

This commit introduces an hopefully cleaner way to handle this case by
defining an user exception 'DBConfigNotFoundError' that is raised if the
db.ini file cannot be read.  This exception is ignored in the early
connection made in scripts/msnoise.py while looking up installed
plugins, as at this stage the file can legitimately not exist.  The
exception is however caught in the run() function of the same file to
present a clean error message to the user before exiting the program.