Add the --get option to the config subcommand
This commit adds the --get option to the 'config' subcommand to allow to
quickly retrieve a configuration value from the database, in a similar
way the --set option allows to change the value.

It also renames the 'set' parameter of the config() function to 'set_'
to avoid overriding the 'set' python keywork.