Use click.Group.command decorators
In the msnoise.py script, we currently use the click.command decorator
to designate all click commands and subcommands, and we maintain a list
of cli.add_command() calls at the end of the file.

This commit instead uses the click.Group.command decorator to designate
click commands and subcommands. It allows to immediately see with which
group a command is linked, and avoids us to maintain the list of
cli.add_command.