Cleanly exit on database connection error
If MSNoise is unable to connect to the database server, a long traceback
is produced to show the error.  This could lead some users to believe
they hit a bug in the code, while the problem is really the database
server being unreachable or unavailable.

This commit catches early database connection errors thrown while trying
to read the list of configured plugins, and outputs a more readable
error message stating the error.