Factorise and rework 'config' and 'info' subcommands
The 'msnoise config show' subcommand recently introduced was duplicating
parts of the output of 'msnoise info'.

This commit changes the 'msnoise config show' to a 'msnoise config get'
that takes one or more mandatory parameter names and display their
value.

It also rework the code generating the output for 'msnoise info' and
splits the info() function into smaller subfunctions for better code
readability.  Slight improvements are brought to this output: mainly
improved column alignments for filter and job information, and a new
function showing the content of the 'db.ini' file.