Rework scan-archive multiprocessing scan
This is an important rework of the code in s01scan_archive.py to make it
more readable, easier to maintain, and more efficient while scanning
files and directories.  (This was first initiated to correct some
multiprocessing-related bug/odd behaviour on CentOS 6.5).

It uses multiprocessing.Pool objects to distribute the scan of the
archive directories among multiple child processes, if the -t/--threads
option was used to indicate the number of concurrent processes to use.

The logging system was also slightly changed in s01scan_archive.py and
scripts/msnoise.py.  A logger instance is now setup through the
api.get_logger() function (and a new instance is created in child
processes to add the process id in the log record format).