New option --crondays to scan-archive
A new option '--crondays' is introduced to the 'scan-archive' subcommand
that overrides the 'crondays' configuration parameter.  It allows to
explicitly change the value of this parameter from the command line.

The value of this parameter can now be positive and still represent a
number of days in the past (msnoise don't handle files modified in the
future yet!).  For backward compatibility, negative values are accepted
and their sign is silently inverted.