Fix documentation generation for click commands
Recent changes in the way the click commands (and subcommands) are
defined in script/msnoise.py broke the autodoc feature of the
documentation generation that could not find the commands any more.

This commit updates the code in conf.py and hopefully makes it a bit
shorter and more pythonic in the same time.