Add new option 'archive_format' and fix forced archive format
The 'scan_archive' functionnality inadvertently forces the 'MSEED'
format when calling the obspy.core.stream.read function (initially for
tests).

This fixes this error and adds a new 'archive_format' configuration
parameter that allows to force the format to avoid the slightly slower
format auto-detection performed by obspy when the specified format is
None (or not recognised, apparently).