Make the 'crondays' parameter more flexible
The 'crondays' was limited to a integer, which was a regression compared
to version 1.5.

These changes allow to use a float to designate a number of days in the
past.  They also allow to use a string to designate an optional number
of weeks, days, and hours in the past in the format 'Xw Xd Xh', where
each group is optional and separated by any number of optional blank
character that match the '\s' pattern of the re module.