Sub-daily CCFs are now default
I've modified the msnoise workflow to allow sub-daily mov_stacks. This means:
* keep_all is now default **Y**
* The stack method will read ALL the CCFs of corr_duratino length and apply: 1/ a rolling mean (for now, can be adapted later to apply other stack methods) of a given duration (e.g. "1d" (d=day in pandas vocabulary) and 2/ a resampling of (possibly) other duration (e.g. "12h").
This results in the STACKS' netcdf files containing 2 CCFs per day, and all subsequent steps have been adapted to work with any number of indices present in the netcdf files.

There might still be some small issues of missing CCFs at the edge of days (or the last window etc), to check.

This introduces a solid backward incompatibility, in the parameters, the code & the file structure.